+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo12_0_0.java
