+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo1_2_0.java
