+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo1_4_0.java
@@ -55,7 +55,8 @@ public class MigrateTo1_4_0 implements Migration {
         }
         ImpersonationConstants.setupImpersonationService(session, realm);
 
-        migrateLDAPMappers(session, realm);
+        // TODO: changed the order of execution here ... does that matter?
+        // migrateLDAPMappers(session, realm);
         migrateUsers(session, realm);
     }
 
@@ -63,19 +64,6 @@ public class MigrateTo1_4_0 implements Migration {
     public void migrateImport(KeycloakSession session, RealmModel realm, RealmRepresentation rep, boolean skipUserDependent) {
         migrateRealm(session, realm);
 
-    }
-
-    private void migrateLDAPMappers(KeycloakSession session, RealmModel realm) {
-        List<String> mandatoryInLdap = Arrays.asList("username", "username-cn", "first name", "last name");
-        realm.getUserStorageProvidersStream()
-                .filter(providerModel -> Objects.equals(providerModel.getProviderId(), LDAPConstants.LDAP_PROVIDER))
-                .forEachOrdered(providerModel -> realm.getComponentsStream(providerModel.getId())
-                        .filter(mapper -> mandatoryInLdap.contains(mapper.getName()))
-                        .forEach(mapper -> {
-                            mapper = new ComponentModel(mapper);  // don't want to modify cache
-                            mapper.getConfig().putSingle("is.mandatory.in.ldap", "true");
-                            realm.updateComponent(mapper);
-                        }));
     }
 
     private void migrateUsers(KeycloakSession session, RealmModel realm) {
