+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo1_6_0.java
