+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo1_7_0.java
