+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo1_8_0.java
@@ -25,6 +25,7 @@ import org.keycloak.models.RealmModel;
 import org.keycloak.models.utils.KeycloakModelUtils;
 import org.keycloak.representations.idm.RealmRepresentation;
 import org.keycloak.storage.UserStorageProviderModel;
+import org.keycloak.storage.UserStorageUtil;
 
 import java.util.Objects;
 
@@ -50,7 +51,7 @@ public class MigrateTo1_8_0 implements Migration {
     }
 
     protected void migrateRealm(RealmModel realm) {
-        realm.getUserStorageProvidersStream()
+        UserStorageUtil.getUserStorageProvidersStream(realm)
                 .filter(fedProvider -> Objects.equals(fedProvider.getProviderId(), LDAPConstants.LDAP_PROVIDER))
                 .filter(this::isActiveDirectory)
                 .filter(fedProvider -> Objects.isNull(getMapperByName(realm, fedProvider, "MSAD account controls")))
