+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo1_9_2.java
