+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo2_0_0.java
