+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo2_2_0.java
