+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo2_5_0.java
