+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo3_0_0.java
