+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo3_2_0.java
