+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo3_4_2.java
