+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo4_2_0.java
