+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo4_6_0.java
