+++ model/legacy-private/src/main/java/org/keycloak/migration/migrators/MigrateTo6_0_0.java
