+++ model/legacy-private/src/main/java/org/keycloak/storage/managers/UserStorageSyncManager.java
@@ -31,6 +31,7 @@ import org.keycloak.models.utils.KeycloakModelUtils;
 import org.keycloak.storage.UserStorageProvider;
 import org.keycloak.storage.UserStorageProviderFactory;
 import org.keycloak.storage.UserStorageProviderModel;
+import org.keycloak.storage.UserStorageUtil;
 import org.keycloak.storage.user.ImportSynchronization;
 import org.keycloak.storage.user.SynchronizationResult;
 import org.keycloak.timer.TimerProvider;
@@ -61,7 +62,7 @@ public class UserStorageSyncManager {
             public void run(KeycloakSession session) {
                 Stream<RealmModel> realms = session.realms().getRealmsWithProviderTypeStream(UserStorageProvider.class);
                 realms.forEach(realm -> {
-                    Stream<UserStorageProviderModel> providers = realm.getUserStorageProvidersStream();
+                    Stream<UserStorageProviderModel> providers = UserStorageUtil.getUserStorageProvidersStream(realm);
                     providers.forEachOrdered(provider -> {
                         UserStorageProviderFactory factory = (UserStorageProviderFactory) session.getKeycloakSessionFactory().getProviderFactory(UserStorageProvider.class, provider.getProviderId());
                         if (factory instanceof ImportSynchronization && provider.isImportEnabled()) {
@@ -165,7 +166,7 @@ public class UserStorageSyncManager {
 
 
     public static void notifyToRefreshPeriodicSyncAll(KeycloakSession session, RealmModel realm, boolean removed) {
-        realm.getUserStorageProvidersStream().forEachOrdered(fedProvider ->
+        UserStorageUtil.getUserStorageProvidersStream(realm).forEachOrdered(fedProvider ->
            notifyToRefreshPeriodicSync(session, realm, fedProvider, removed));
     }
     
@@ -267,7 +268,7 @@ public class UserStorageSyncManager {
             @Override
             public void run(KeycloakSession session) {
                 RealmModel persistentRealm = session.realms().getRealm(realmId);
-                persistentRealm.getUserStorageProvidersStream()
+                UserStorageUtil.getUserStorageProvidersStream(persistentRealm)
                         .filter(persistentFedProvider -> Objects.equals(provider.getId(), persistentFedProvider.getId()))
                         .forEachOrdered(persistentFedProvider -> {
                             // Update persistent provider in DB
