+++ server-spi-private/src/main/java/org/keycloak/storage/ExportImportManager.java
@@ -1,0 +1,11 @@
+package org.keycloak.storage;
+
+import org.keycloak.models.RealmModel;
+import org.keycloak.representations.idm.RealmRepresentation;
+
+/**
+ * @author Alexander Schwartz
+ */
+public interface ExportImportManager {
+    void importRealm(RealmRepresentation rep, RealmModel newRealm);
+}
