+++ server-spi-private/src/main/java/org/keycloak/storage/MigrationManager.java
@@ -1,0 +1,14 @@
+package org.keycloak.storage;
+
+import org.keycloak.models.RealmModel;
+import org.keycloak.representations.idm.RealmRepresentation;
+
+/**
+ * @author Alexander Schwartz
+ */
+public interface MigrationManager {
+
+    void migrate();
+
+    void migrate(RealmModel realm, RealmRepresentation rep, boolean skipUserDependent);
+}
