+++ server-spi/src/main/java/org/keycloak/models/RealmModel.java
@@ -23,8 +23,6 @@ import org.keycloak.component.ComponentModel;
 import org.keycloak.provider.Provider;
 import org.keycloak.provider.ProviderEvent;
 import org.keycloak.storage.SearchableModelField;
-import org.keycloak.storage.UserStorageProvider;
-import org.keycloak.storage.UserStorageProviderModel;
 import org.keycloak.storage.client.ClientStorageProvider;
 import org.keycloak.storage.client.ClientStorageProviderModel;
 import org.keycloak.storage.role.RoleStorageProvider;
@@ -661,25 +659,6 @@ public interface RealmModel extends RoleContainerModel {
     Stream<ComponentModel> getComponentsStream();
 
     ComponentModel getComponent(String id);
-
-    /**
-     * @deprecated Use {@link #getUserStorageProvidersStream() getUserStorageProvidersStream} instead.
-     */
-    @Deprecated
-    default List<UserStorageProviderModel> getUserStorageProviders() {
-        return getUserStorageProvidersStream().collect(Collectors.toList());
-    }
-
-    /**
-     * Returns sorted {@link UserStorageProviderModel UserStorageProviderModel} as a stream.
-     * It should be used with forEachOrdered if the ordering is required.
-     * @return Sorted stream of {@link UserStorageProviderModel}. Never returns {@code null}.
-     */
-    default Stream<UserStorageProviderModel> getUserStorageProvidersStream() {
-        return getComponentsStream(getId(), UserStorageProvider.class.getName())
-                .map(UserStorageProviderModel::new)
-                .sorted(UserStorageProviderModel.comparator);
-    }
 
     /**
      * @deprecated Use {@link #getClientStorageProvidersStream() getClientStorageProvidersStream} instead.
